Tests for New-BTAppId + only run whatif on TF_Build

+++ Tests/BurntToast.Tests.ps1
@@ -10,19 +10,21 @@ Describe 'New-BTAction' {
             New-BTAction | Should BeOfType 'Microsoft.Toolkit.Uwp.Notifications.ToastActionsCustom'
         }
 
-        Start-Transcript tmp.log
-        try {
-            New-BTAction -WhatIf
-        }
-        finally {
-            Stop-Transcript
-            $Log = (Get-Content tmp.log).Where({ $_ -match "What if: " })
-            Remove-Item tmp.log
-        }
+        if ($env:TF_BUILD) {
+            Start-Transcript tmp.log
+            try {
+                New-BTAction -WhatIf
+            }
+            finally {
+                Stop-Transcript
+                $Log = (Get-Content tmp.log).Where({ $_ -match "What if: " })
+                Remove-Item tmp.log
+            }
 
-        It 'has consitent WhatIf response' {
-            $Expected = 'What if: Performing the operation "New-BTAction" on target "returning: [ToastActionsCustom] with 0 Inputs, 0 Buttons, and 0 ContextMenuItems".'
-            $Log -join [System.Environment]::NewLine | Should Be $Expected
+            It 'has consitent WhatIf response' {
+                $Expected = 'What if: Performing the operation "New-BTAction" on target "returning: [ToastActionsCustom] with 0 Inputs, 0 Buttons, and 0 ContextMenuItems".'
+                $Log -join [System.Environment]::NewLine | Should Be $Expected
+            }
         }
     }
     Context 'snooze and dismiss modal' {
@@ -34,19 +36,21 @@ Describe 'New-BTAction' {
             New-BTAction -SnoozeAndDismiss | Should BeOfType 'Microsoft.Toolkit.Uwp.Notifications.ToastActionsSnoozeAndDismiss'
         }
 
-        Start-Transcript tmp.log
-        try {
-            New-BTAction -SnoozeAndDismiss -WhatIf
-        }
-        finally {
-            Stop-Transcript
-            $Log = (Get-Content tmp.log).Where({ $_ -match "What if: " })
-            Remove-Item tmp.log
-        }
+        if ($env:TF_BUILD) {
+            Start-Transcript tmp.log
+            try {
+                New-BTAction -SnoozeAndDismiss -WhatIf
+            }
+            finally {
+                Stop-Transcript
+                $Log = (Get-Content tmp.log).Where({ $_ -match "What if: " })
+                Remove-Item tmp.log
+            }
 
-        It 'has consitent WhatIf response' {
-            $Expected = 'What if: Performing the operation "New-BTAction" on target "returning: [ToastActionsSnoozeAndDismiss] with 0 Inputs, 0 Buttons, and 0 ContextMenuItems".'
-            $Log -join [System.Environment]::NewLine | Should Be $Expected
+            It 'has consitent WhatIf response' {
+                $Expected = 'What if: Performing the operation "New-BTAction" on target "returning: [ToastActionsSnoozeAndDismiss] with 0 Inputs, 0 Buttons, and 0 ContextMenuItems".'
+                $Log -join [System.Environment]::NewLine | Should Be $Expected
+            }
         }
     }
     Context 'single clickable button' {
@@ -58,19 +62,21 @@ Describe 'New-BTAction' {
             New-BTAction -Buttons (New-BTButton -Content 'Google' -Arguments 'https://google.com') | Should BeOfType 'Microsoft.Toolkit.Uwp.Notifications.ToastActionsCustom'
         }
 
-        Start-Transcript tmp.log
-        try {
-            New-BTAction -Buttons (New-BTButton -Content 'Google' -Arguments 'https://google.com') -WhatIf
-        }
-        finally {
-            Stop-Transcript
-            $Log = (Get-Content tmp.log).Where({ $_ -match "What if: " })
-            Remove-Item tmp.log
-        }
+        if ($env:TF_BUILD) {
+            Start-Transcript tmp.log
+            try {
+                New-BTAction -Buttons (New-BTButton -Content 'Google' -Arguments 'https://google.com') -WhatIf
+            }
+            finally {
+                Stop-Transcript
+                $Log = (Get-Content tmp.log).Where({ $_ -match "What if: " })
+                Remove-Item tmp.log
+            }
 
-        It 'has consitent WhatIf response' {
-            $Expected = 'What if: Performing the operation "New-BTAction" on target "returning: [ToastActionsCustom] with 0 Inputs, 1 Buttons, and 0 ContextMenuItems".'
-            $Log -join [System.Environment]::NewLine | Should Be $Expected
+            It 'has consitent WhatIf response' {
+                $Expected = 'What if: Performing the operation "New-BTAction" on target "returning: [ToastActionsCustom] with 0 Inputs, 1 Buttons, and 0 ContextMenuItems".'
+                $Log -join [System.Environment]::NewLine | Should Be $Expected
+            }
         }
     }
     Context 'mixed content: button & context menu' {
@@ -87,22 +93,95 @@ Describe 'New-BTAction' {
             New-BTAction -Buttons $Button -ContextMenuItems $ContextMenuItem | Should BeOfType 'Microsoft.Toolkit.Uwp.Notifications.ToastActionsCustom'
         }
 
-        Start-Transcript tmp.log
-        try {
-            New-BTAction -Buttons $Button -ContextMenuItems $ContextMenuItem -WhatIf
-        }
-        finally {
-            Stop-Transcript
-            $Log = (Get-Content tmp.log).Where({ $_ -match "What if: " })
-            Remove-Item tmp.log
-        }
+        if ($env:TF_BUILD) {
+            Start-Transcript tmp.log
+            try {
+                New-BTAction -Buttons $Button -ContextMenuItems $ContextMenuItem -WhatIf
+            }
+            finally {
+                Stop-Transcript
+                $Log = (Get-Content tmp.log).Where({ $_ -match "What if: " })
+                Remove-Item tmp.log
+            }
 
-        It 'has consitent WhatIf response' {
-            $Expected = 'What if: Performing the operation "New-BTAction" on target "returning: [ToastActionsCustom] with 0 Inputs, 1 Buttons, and 1 ContextMenuItems".'
-            $Log -join [System.Environment]::NewLine | Should Be $Expected
+            It 'has consitent WhatIf response' {
+                $Expected = 'What if: Performing the operation "New-BTAction" on target "returning: [ToastActionsCustom] with 0 Inputs, 1 Buttons, and 1 ContextMenuItems".'
+                $Log -join [System.Environment]::NewLine | Should Be $Expected
+            }
         }
     }
 }
+
+Describe 'New-BTAppId' {
+    Mock New-Item {}
+    Mock New-ItemProperty {}
+
+    Context 'running without arguments' {
+        It 'runs without errors' {
+            { New-BTAppId } | Should Not Throw
+        }
+
+        It 'attempts to add item to the registry' {
+            Mock Test-Path {return $false}
+
+            New-BTAppId
+
+            Assert-MockCalled New-Item -Exactly 1 -Scope It
+            Assert-MockCalled New-ItemProperty -Exactly 1 -Scope It
+        }
+
+        if ($env:TF_BUILD) {
+            Start-Transcript tmp.log
+            try {
+                New-BTAppId -WhatIf
+            }
+            finally {
+                Stop-Transcript
+                $Log = (Get-Content tmp.log).Where({ $_ -match "What if: " })
+                Remove-Item tmp.log
+            }
+
+            It 'has consitent WhatIf response' {
+                $Expected = 'What if: Performing the operation "New-BTAppId" on target "creating: ''HKCU:\SOFTWARE\Microsoft\Windows\CurrentVersion\Notifications\Settings\{1AC14E77-02E7-4E5D-B744-2EB1AE5198B7}\WindowsPowerShell\v1.0\powershell.exe'' with property ''ShowInActionCenter'' set to ''1'' (DWORD)".'
+                $Log -join [System.Environment]::NewLine | Should Be $Expected
+            }
+        }
+    }
+    Context 'running with custom AppId' {
+        It 'runs without errors' {
+            { New-BTAppId -AppId 'Script Checker' } | Should Not Throw
+        }
+
+        It 'returns a [ToastActionsSnoozeAndDismiss] object' {
+            New-BTAppId -AppId 'Script Checker'
+
+            Assert-MockCalled New-Item -Exactly 1 -Scope It
+            Assert-MockCalled New-ItemProperty -Exactly 1 -Scope It
+        }
+
+        if ($env:TF_BUILD) {
+            Start-Transcript tmp.log
+            try {
+                New-BTAppId -AppId 'Script Checker'
+            }
+            finally {
+                Stop-Transcript
+                $Log = (Get-Content tmp.log).Where({ $_ -match "What if: " })
+                Remove-Item tmp.log
+            }
+
+            It 'has consitent WhatIf response' {
+                $Expected = 'What if: Performing the operation "New-BTAppId" on target "creating: ''HKCU:\SOFTWARE\Microsoft\Windows\CurrentVersion\Notifications\Settings\Script Checker'' with property ''ShowInActionCenter'' set to ''1'' (DWORD)".'
+                $Log -join [System.Environment]::NewLine | Should Be $Expected
+            }
+        }
+    }
+}
+
+
+
+
+
 Describe 'New-BurntToastNotification' {
     Context 'running without arguments' {
         It 'runs without errors' {
